Changes to key functionality for the FFgame
So I tried adding some keyboard functionality at the end of the FF game so you can select the "fix it" and "its ok" options with the arrow keys. It highlights the options and once selected continues to the main game.

Also I found that the game ended after only inspecting 3 of the sprites out of the 6 and then the game abruptly ends. I found that a little annoying so I also implemented code so that the game does not end until you answer all 6 questions through the sprites.